+++ src/Organizations/OrganizationsList/OrganizationsListSearchConfig.js
@@ -12,7 +12,11 @@ export const searchableIndexes = [
     labelId: 'ui-organizations.search.keyword',
     value: '',
   },
-  ...indexes.map(index => ({ labelId: `ui-organizations.search.${index}`, value: index })),
+  ...indexes.map(index => ({
+    labelId: `ui-organizations.search.${index}`,
+    placeholderId: index === 'language' && 'ui-organizations.search.placeholder.language',
+    value: index,
+  })),
 ];
 
 export const getKeywordQuery = query => indexes.reduce(